Removed DockPanelSkinBuilder class. Fixed color, text font, and width.
