+++ WinFormsUI/Docking/VS2005Theme.cs
@@ -1,5 +1,6 @@
 using System;
-using WeifenLuo.WinFormsUI.Docking.Skins;
+using System.Drawing;
+using System.Drawing.Drawing2D;
 
 namespace WeifenLuo.WinFormsUI.Docking
 {
@@ -18,7 +19,7 @@ namespace WeifenLuo.WinFormsUI.Docking
             {
                 throw new NullReferenceException("dockPanel");
             }
-            
+
             Measures.SplitterSize = 4;
             dockPanel.Extender.DockPaneCaptionFactory = null;
             dockPanel.Extender.AutoHideStripFactory = null;
@@ -26,7 +27,45 @@ namespace WeifenLuo.WinFormsUI.Docking
             dockPanel.Extender.DockPaneStripFactory = null;
             dockPanel.Extender.DockPaneSplitterControlFactory = null;
             dockPanel.Extender.DockWindowFactory = null;
-            dockPanel.SkinStyle = Style.VisualStudio2005;
+            dockPanel.Skin = CreateVisualStudio2005();
+        }
+
+        internal static DockPanelSkin CreateVisualStudio2005()
+        {
+            DockPanelSkin skin = new DockPanelSkin();
+
+            skin.AutoHideStripSkin.DockStripGradient.StartColor = SystemColors.ControlLight;
+            skin.AutoHideStripSkin.DockStripGradient.EndColor = SystemColors.ControlLight;
+            skin.AutoHideStripSkin.TabGradient.TextColor = SystemColors.ControlDarkDark;
+
+            skin.DockPaneStripSkin.DocumentGradient.DockStripGradient.StartColor = SystemColors.Control;
+            skin.DockPaneStripSkin.DocumentGradient.DockStripGradient.EndColor = SystemColors.Control;
+            skin.DockPaneStripSkin.DocumentGradient.ActiveTabGradient.StartColor = SystemColors.ControlLightLight;
+            skin.DockPaneStripSkin.DocumentGradient.ActiveTabGradient.EndColor = SystemColors.ControlLightLight;
+            skin.DockPaneStripSkin.DocumentGradient.InactiveTabGradient.StartColor = SystemColors.ControlLight;
+            skin.DockPaneStripSkin.DocumentGradient.InactiveTabGradient.EndColor = SystemColors.ControlLight;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.DockStripGradient.StartColor = SystemColors.ControlLight;
+            skin.DockPaneStripSkin.ToolWindowGradient.DockStripGradient.EndColor = SystemColors.ControlLight;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveTabGradient.StartColor = SystemColors.Control;
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveTabGradient.EndColor = SystemColors.Control;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveTabGradient.StartColor = Color.Transparent;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveTabGradient.EndColor = Color.Transparent;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveTabGradient.TextColor = SystemColors.ControlDarkDark;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveCaptionGradient.StartColor = SystemColors.GradientActiveCaption;
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveCaptionGradient.EndColor = SystemColors.ActiveCaption;
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveCaptionGradient.LinearGradientMode = LinearGradientMode.Vertical;
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveCaptionGradient.TextColor = SystemColors.ActiveCaptionText;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveCaptionGradient.StartColor = SystemColors.GradientInactiveCaption;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveCaptionGradient.EndColor = SystemColors.InactiveCaption;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveCaptionGradient.LinearGradientMode = LinearGradientMode.Vertical;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveCaptionGradient.TextColor = SystemColors.InactiveCaptionText;
+
+            return skin;
         }
     }
 }
