+++ WinFormsUI/Docking/VS2012LightTheme.cs
@@ -1,5 +1,6 @@
 using System;
-using WeifenLuo.WinFormsUI.Docking.Skins;
+using System.Drawing;
+using System.Drawing.Drawing2D;
 
 namespace WeifenLuo.WinFormsUI.Docking
 {
@@ -26,7 +27,7 @@ namespace WeifenLuo.WinFormsUI.Docking
             dockPanel.Extender.DockPaneStripFactory = new VS2012LightDockPaneStripFactory();
             dockPanel.Extender.DockPaneSplitterControlFactory = new VS2012LightDockPaneSplitterControlFactory();
             dockPanel.Extender.DockWindowFactory = new VS2012LightDockWindowFactory();
-            dockPanel.SkinStyle = Style.VisualStudio2012Light;
+            dockPanel.Skin = CreateVisualStudio2012Light();
         }
 
         private class VS2012LightAutoHideWindowFactory : DockPanelExtender.IAutoHideWindowFactory
@@ -75,6 +76,50 @@ namespace WeifenLuo.WinFormsUI.Docking
             {
                 return new VS2012LightDockWindow(dockPanel, dockState);
             }
+        }
+
+        public static DockPanelSkin CreateVisualStudio2012Light()
+        {
+            var specialBlue = Color.FromArgb(0xFF, 0x00, 0x7A, 0xCC);
+            var activeTab = Color.FromArgb(0xFF, 176, 203, 241);
+            var inactiveTab = Color.FromArgb(0xFF, 141, 163, 193);
+            var skin = new DockPanelSkin();
+            
+            skin.AutoHideStripSkin.DockStripGradient.StartColor = specialBlue;
+            skin.AutoHideStripSkin.DockStripGradient.EndColor = SystemColors.ControlLight;
+            skin.AutoHideStripSkin.TabGradient.TextColor = SystemColors.ControlDarkDark;
+
+            skin.DockPaneStripSkin.DocumentGradient.DockStripGradient.StartColor = SystemColors.Control;
+            skin.DockPaneStripSkin.DocumentGradient.DockStripGradient.EndColor = SystemColors.Control;
+            skin.DockPaneStripSkin.DocumentGradient.ActiveTabGradient.StartColor = activeTab;
+            skin.DockPaneStripSkin.DocumentGradient.ActiveTabGradient.EndColor = activeTab;
+            skin.DockPaneStripSkin.DocumentGradient.ActiveTabGradient.TextColor = Color.Black;
+            skin.DockPaneStripSkin.DocumentGradient.InactiveTabGradient.StartColor = inactiveTab;
+            skin.DockPaneStripSkin.DocumentGradient.InactiveTabGradient.EndColor = inactiveTab;
+            skin.DockPaneStripSkin.DocumentGradient.InactiveTabGradient.TextColor = Color.Black;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.DockStripGradient.StartColor = SystemColors.Control;
+            skin.DockPaneStripSkin.ToolWindowGradient.DockStripGradient.EndColor = SystemColors.Control;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveTabGradient.StartColor = SystemColors.ControlLightLight;
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveTabGradient.EndColor = SystemColors.ControlLightLight;
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveTabGradient.TextColor = specialBlue;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveTabGradient.StartColor = SystemColors.Control;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveTabGradient.EndColor = SystemColors.Control;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveTabGradient.TextColor = SystemColors.GrayText;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveCaptionGradient.StartColor = specialBlue;
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveCaptionGradient.EndColor = Color.FromArgb(80, 170, 220); // dots in caption
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveCaptionGradient.LinearGradientMode = LinearGradientMode.Vertical;
+            skin.DockPaneStripSkin.ToolWindowGradient.ActiveCaptionGradient.TextColor = Color.White;
+
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveCaptionGradient.StartColor = SystemColors.Control;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveCaptionGradient.EndColor = SystemColors.ControlDark;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveCaptionGradient.LinearGradientMode = LinearGradientMode.Vertical;
+            skin.DockPaneStripSkin.ToolWindowGradient.InactiveCaptionGradient.TextColor = SystemColors.GrayText;
+
+            return skin;
         }
     }
 }
